Removes unnecessary (?^:) wrapper around paren regex. Bumps req to 5.10
The (?^ cluster is causing failures on pre-5.14 version of perl. It does
not appear to be necessary given our use of the regex. On the other
hand, the use of (?-1) was added in 5.10, so let's bump the perl
requirement back to 5.10 (it is hard to imagine anyone using modern
method syntax on a pre-5.10 perl).